refactor(cli): extract business logic to domain modules (#611) — v1.3.73
Closes #611 (#arch-h8).

cli.py was 1,395 LOC, much of which was domain logic that shouldn't
live in a CLI shim. This PR moves two non-trivial blocks out:

1. synthesize_estimate_report → llmwiki/synth/estimate.py
   The G-07 / #293 incremental-vs-full-force cost-model walk now
   sits next to the rest of the synth pipeline.

2. _adapter_status → llmwiki/adapters/status.py (renamed
   adapter_status — no leading underscore at the new path).
   The configured / will_fire label decision is adapter-domain
   logic; cli.py only needs to print the result.

cli.py re-exports both names so the existing import paths
(from llmwiki.cli import synthesize_estimate_report,
 from llmwiki.cli import _adapter_status) keep working unchanged.

cli.py is now 1,234 LOC (-161). Still not a one-liner, but closer
to argparse-setup + dispatch.

Pure refactor — no behavior change. Full pytest suite passes.
